requirements.txt
mysql_connector_repackaged==0.3.1
tabulate==0.9.0
